# Improve all join performance by append `RowRefList` or `RowRef` to AddedColumns for lazy output