… src/Interpreters/TableJoin.h
… tests/performance/all_join_opt.xml