@KevinyhZou 2024-05-17
![image](https://github.com/ClickHouse/ClickHouse/assets/37431499/cb648b8b-e577-4b00-b952-6bb27f9ead45)

the performance improves in join_in_memorys