@nickitat 2024-06-13
pls check reported slowdowns in `storage_join_direct_join`