@KevinyhZou 2024-06-14
yes, I see. Maybe it it same problem as `jon_append_block`, has the cache miss problems in the buildOutput for loop.