@KevinyhZou 2024-06-17
It seems the reference in the for loop cause `storage_join_direct_join` slowdown,  and it is fixed now. But still has 9.7% performance slowdown in `jon_append_block`.