@nickitat 2024-06-18
overall looks ok for me, I'll check perf myself