@KevinyhZou 2024-07-16
@nickitat I had add a settings `join_output_by_rowlist_perkey_rows_threshold` to control not to use output by row ref list, if the average rows perkey in right table is below this threshold,  and now it has no performance slowdown.  could you help to review this?